Worker/requirements.txt
python-dotenv
requests 
GPUtil